zemberek-veriisleme/requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
joblib==0.14.1
JPype1==0.7.2
numpy==1.18.2
openpyxl==3.0.3
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
xlrd==1.2.0
